(NOT WORKING) refactor into contexts
