--- src/context/index.jsx
@@ -1,4 +1,0 @@
-export * from './mesh';
-export * from './api';
-export * from './contract';
-export {sceneService} from './scene';
