--- src/library/bones.js
@@ -1,74 +1,0 @@
-export default [   
-    {
-        "name": "Hip",
-        "bone": "Torso_Hip"
-    },
-    {
-        "name": "Spine",
-        "bone": "Torso_Spine"
-    },
-    {
-        "name": "Chest",
-        "bone": "Torso_Chest"
-    },
-    {   
-        "name": "Neck",
-        "bone": "Torso_Neck"
-    },
-    {
-        "name": "Left Shoulder",
-        "bone": "Torso_Sholder_L"
-    },
-    {
-        "name": "Left Upper Arm",
-        "bone": "Torso_UpperArm_L"
-    },
-    {
-        "name": "Left Lower Arm",
-        "bone": "ArmL_LowerArm_L"
-    },
-    {
-        "name": "Left Hand",
-        "bone": "ArmL_Hand_L"
-    },
-    {
-        "name": "Right Shoulder",
-        "bone": "Torso_Sholder_R"
-    },
-    {
-        "name": "Right Upper Arm",
-        "bone": "Torso_UpperArm_R"
-    },
-    {
-        "name": "Right Lower Arm",
-        "bone": "ArmR_LowerArm_R"
-    },
-    {
-        "name": "Right Hand",
-        "bone": "ArmR_Hand_R"
-    },
-    {
-        "name": "Left Upper Leg",
-        "bone": "Torso_UpperLeg_L"
-    },
-    {
-        "name": "Left Lower Leg",
-        "bone": "LegL_LowerLeg_L"
-    },
-    {
-        "name": "Left Foot",
-        "bone": "LegL_Foot_L"
-    },
-    {
-        "name": "Right Upper Leg",
-        "bone": "Torso_UpperLeg_R"
-    },
-    {
-        "name": "Right Lower Leg",
-        "bone": "LegR_LowerLeg_R"
-    },
-    {
-        "name": "Right Foot",
-        "bone": "LegR_Foot_R"
-    }
-]
